+++ src/wfg/native_ui/internal/ui/dialog/ModalInterceptor.java
@@ -1,4 +1,4 @@
-package wfg.native_ui.internal.ui.functional;
+package wfg.native_ui.internal.ui.dialog;
 
 import static wfg.native_ui.util.UIConstants.screenH;
 import static wfg.native_ui.util.UIConstants.screenW;
@@ -8,7 +8,6 @@ import java.util.List;
 import com.fs.starfarer.api.input.InputEventAPI;
 import com.fs.starfarer.api.ui.UIPanelAPI;
 
-import wfg.native_ui.internal.ui.dialog.ModalDialog;
 import wfg.native_ui.ui.panel.CustomPanel;
 
 public class ModalInterceptor extends CustomPanel<ModalInterceptor> {
@@ -30,5 +29,14 @@ public class ModalInterceptor extends CustomPanel<ModalInterceptor> {
 
             e.consume();
         }
+    }
+
+    @Override
+    public void advance(float delta) {
+        super.advance(delta);
+
+        if (dialog.isBeingDismissed() && !dialog.suspendEventInterception &&
+            dialog.getFaderBrightness() < 0.5f
+        ) { m_parent.removeComponent(m_panel); }
     }
 }