+++ src/wfg/native_ui/ui/core/UIElementAPI.java
@@ -31,4 +31,5 @@ public interface UIElementAPI extends UIComponentAPI {
 
     void bringToFront();
     void sendToBack();
+    void detach();
 }